Making Ribs vector more efficient
This commit is part way through making the ribs vector more efficient. Before this commit, any time the player threw a rib, a rib was added to the vector, and so the vector's size was increased. Any time a rib was picked up, it was removed from the vector, reducing the vector's size. The constant changing in size is inefficient.

Now, the ProjectileManager::init() function adds 6 ribs to the vector, which all spawn off screen with zero velocity.

In the game's current state, the ProjectileManager::processInput() function applies movement to the rib in the current ammo level's position in the ribs vector. This does seem to have introduced an issue where a rib can go missing in game, even when the vector size is still 6. This has been seen twice, but cannot be replicated on purpose.

The remove_if function is now only used as a backup. When a rib is picked up, its position is now moved off the screen, instead of having it's distance value set to 1001, and so removed from the vector.

The Player::clean() function has also been updated to include setting variables to nullptr's, and deleting them afterwards.