Ribs vector efficient + bug fix
The bug described in the previous commit occurred when ribs were not picked up in the reverse order that they were thrown, causing ribs to be "lost". A new bool variable has been added to the Rib struct to define whether an individual rib is still in the Player's body, or if it has been thrown. When the player throws a rib, the program finds the first rib in the vector where isThrowable == true and applies values to its variables in order to throw it. These changes mean that the size of the Ribs vector never changes, while allowing ribs to be thrown and picked up.
Similar efficiencies still need to be applied to the zombies vector.

Player and Zombie colliders related to tiles have been changed so that the upper third of the sprites do not collide, providing more space in the map.