Main Menu Foundations (needs re-work)
There is an incredibly basic main menu which has one button to take you to the game. The timer in the game starts from the start of the program, not after the player leaves the menu.

To create an efficient (and simpler) main menu, the program needs to be re-worked in terms of where it's functionality is. Currently all functionality feeds into the GameLoop class. I plan on changing this so that there is a "Levels" class which will be the base for everything. Then the menu class would be used for menus, with its functionality completely separate to the GameLoop class.