Zombie efficiencies, rotating ribs and more
Zombies are now more efficient, similar to how ribs were changed in the 12/07/2023 commit. This means the vector which holds the zombie instances does not decrease and increase as zombies are defeated and respawned. Instead, defeated zombies are just moved off the screen, and once they respawn their co-ordinates are randomised again.

The bug where zombies would walk out of the map has finally been fixed. I still believe this to be a result of the zombie colliding with more than one wall tile per frame, causing it to turn around more than once. A bool variable has been added to determine if the zombie has already hit a wall in the current frame, and stops checking if it is true. This means the collisions map has been simplified to just zeros and ones as originally intended.

Ribs also now rotate after being thrown, until they hit an object. The original "rotation" variable has been changed to throwRotation, which determines the angle that the rib is thrown at. A spinRotation variable has been added, which determines the angle that the rib is rendered at, allowing it to spin after being thrown.

A lot of code has been cleaned up as well, including removing debug outputs, moving code around to only run when required, and general tidying